dplane_fpm_sonic: carry 16-bit member weight through NHG hashing

+++ src/sonic-frr/dplane_fpm_sonic/fpm_nhg.h
@@ -42,7 +42,7 @@ struct fpm_dplane_nhg;
 
 struct fpm_nhg_child {
 	struct fpm_dplane_nhg *obj;
-	uint8_t weight;
+	uint16_t weight;
 };
 
 struct fpm_dplane_nhg {
